+++ changelog.md
@@ -1,1 +1,1 @@
-- Updated to 1.16.5
+- Fixed storage barrel dropping overstacked items